Refactor HAR interfaces: move to types module and clean up parser
